完成部门计划

+++ src/main/java/com/examw/oa/service/plan/impl/DeptPlanServiceImpl.java
@@ -1,8 +1,12 @@
 package com.examw.oa.service.plan.impl;
 
+import java.util.Comparator;
 import java.util.Date;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.TreeSet;
 import java.util.UUID;
 
 import org.apache.log4j.Logger;
@@ -13,7 +17,6 @@ import com.examw.oa.dao.check.IEntryDao;
 import com.examw.oa.dao.org.IDepartmentDao;
 import com.examw.oa.dao.org.IEmployeeDao;
 import com.examw.oa.dao.plan.IDeptPlanDao;
-import com.examw.oa.dao.plan.IDeptPlanMenberDao;
 import com.examw.oa.domain.check.Entry;
 import com.examw.oa.domain.org.Department;
 import com.examw.oa.domain.org.Employee;
@@ -32,7 +35,6 @@ public class DeptPlanServiceImpl extends BaseDataServiceImpl<DeptPlan, DeptPlanI
 	private static final Logger logger = Logger.getLogger(DeptPlanServiceImpl.class);
 	private IDeptPlanDao deptPlanDao;
 	private IDepartmentDao departmentDao;
-	private IDeptPlanMenberDao deptPlanMenberDao;
 	private IEntryDao entryDao;
 	private IEmployeeDao employeeDao;
 	private Map<Integer,String> typeMap,statusMap;
@@ -52,14 +54,6 @@ public class DeptPlanServiceImpl extends BaseDataServiceImpl<DeptPlan, DeptPlanI
 	 */
 	public void setEmployeeDao(IEmployeeDao employeeDao) {
 		this.employeeDao = employeeDao;
-	}
-	/**
-	 * 设置部门计划成员数据接口。
-	 * @param deptPlanMenberDao
-	 * 部门计划成员数据接口。
-	 */
-	public void setDeptPlanMenberDao(IDeptPlanMenberDao deptPlanMenberDao) {
-		this.deptPlanMenberDao = deptPlanMenberDao;
 	}
 	/**
 	 * 设置奖惩条目数据接口。
@@ -121,13 +115,46 @@ public class DeptPlanServiceImpl extends BaseDataServiceImpl<DeptPlan, DeptPlanI
 	protected DeptPlanInfo changeModel(DeptPlan data) {
 		if(logger.isDebugEnabled()) logger.debug("开始类型转换...");
 		DeptPlanInfo info = new DeptPlanInfo();
-		BeanUtils.copyProperties(data, info);
+		BeanUtils.copyProperties(data,info,new String[]{"members"});
 		if(data.getDepartment() != null){
 			info.setDeptId(data.getDepartment().getId());
 			info.setDeptName(data.getDepartment().getName());
 		}
 		info.setTypeName(this.loadTypeName(info.getType()));
 		info.setStatusName(this.loadStatusName(info.getStatus()));
+		if(data.getMembers() != null && data.getMembers().size() > 0){
+			Set<DeptPlanMemberInfo> members = new TreeSet<DeptPlanMemberInfo>(new Comparator<DeptPlanMemberInfo>(){
+				@Override
+				public int compare(DeptPlanMemberInfo o1, DeptPlanMemberInfo o2) {
+					return (int)(o1.getStartTime().getTime() - o2.getStartTime().getTime());
+				}
+			});
+			for(DeptPlanMember member : data.getMembers()){
+				DeptPlanMemberInfo m = this.changeModel(member);
+				if(m != null)members.add(m);
+			}
+			info.setMembers(members);
+		}
+		return info;
+	}
+	/*
+	 * 部门计划类型转换。
+	 */
+	private DeptPlanMemberInfo changeModel(DeptPlanMember member){
+		if(member == null)return null;
+		DeptPlanMemberInfo info = new DeptPlanMemberInfo();
+		BeanUtils.copyProperties(member, info);
+		if(member.getEmployee() != null){
+			info.setEmpId(member.getEmployee().getId());
+			info.setEmpName(member.getEmployee().getName());
+		}
+		if(member.getEntry() != null){
+			info.setEntryId(member.getEntry().getId());
+			info.setEntryName(member.getEntry().getName());
+		}
+		if(member.getPlan() != null){
+			info.setPlanId(member.getPlan().getId());
+		}
 		return info;
 	}
 	/*
@@ -141,20 +168,45 @@ public class DeptPlanServiceImpl extends BaseDataServiceImpl<DeptPlan, DeptPlanI
 		Boolean isAdded = false;
 		DeptPlan  data = StringUtils.isEmpty(info.getId()) ? null : this.deptPlanDao.load(DeptPlan.class, info.getId());
 		if(isAdded = (data == null)){
-			if(StringUtils.isEmpty(info.getId())){
-				info.setId(UUID.randomUUID().toString());
-				info.setCreateTime(new Date());
-			}
+			if(StringUtils.isEmpty(info.getId()))info.setId(UUID.randomUUID().toString());
+			info.setCreateTime(new Date()); 
 			data = new DeptPlan();
 		}
 		if(!isAdded)info.setCreateTime(data.getCreateTime());
-		BeanUtils.copyProperties(info, data);
+		info.setLastTime(new Date());
+		info.setStatus(DeptPlan.STATUS_NONE);
+		BeanUtils.copyProperties(info, data,new String[]{"members"});
 		if(!StringUtils.isEmpty(info.getDeptId()) && (data.getDepartment() == null || !data.getDepartment().getId().equalsIgnoreCase(info.getDeptId()))){
 			Department d = this.departmentDao.load(Department.class, info.getDeptId());
 			if(d != null) data.setDepartment(d);
 		}
 		if(data.getDepartment() != null) info.setDeptName(data.getDepartment().getName());
 		info.setTypeName(this.loadTypeName(data.getType()));
+		//members
+		Set<DeptPlanMember> members = new HashSet<>();
+		if(info.getMembers() != null && info.getMembers().size() > 0){
+			for(DeptPlanMemberInfo dpi : info.getMembers()){
+				DeptPlanMember dpm = new DeptPlanMember();
+				dpm.setPlan(data);
+				if(StringUtils.isEmpty(dpi.getId())){
+					dpm.setId(UUID.randomUUID().toString());
+				}
+				if(dpi.getCreateTime() == null){
+					dpm.setCreateTime(new Date());
+				}
+				if(!StringUtils.isEmpty(dpi.getEmpId())){
+					Employee employee = this.employeeDao.load(Employee.class,dpi.getEmpId());
+					dpm.setEmployee(employee);
+				}
+				if(!StringUtils.isEmpty(dpi.getEntryId())){
+					Entry entry=this.entryDao.load(Entry.class,dpi.getEntryId());
+					dpm.setEntry(entry);
+				}
+				BeanUtils.copyProperties(dpi, dpm,new String[]{"remarks"});
+				members.add(dpm);
+			}
+		}
+		data.setMembers(members);
 		if(isAdded)this.deptPlanDao.save(data);
 		return info;
 	}
@@ -169,8 +221,13 @@ public class DeptPlanServiceImpl extends BaseDataServiceImpl<DeptPlan, DeptPlanI
 		for(int i = 0; i < ids.length;i++){
 			DeptPlan data = this.deptPlanDao.load(DeptPlan.class, ids[i]);
 			if(data != null){
+				if(data.getStatus() != null && data.getStatus() != DeptPlan.STATUS_NONE){
+					if(logger.isDebugEnabled()) logger.debug("不允许删除数据［"+ ids[i]+"］");
+					continue;
+				}
 				if(logger.isDebugEnabled()) logger.debug("删除数据［"+ ids[i]+"］");
 				this.deptPlanDao.delete(data);
+				 
 			}
 		}
 	}
@@ -193,41 +250,5 @@ public class DeptPlanServiceImpl extends BaseDataServiceImpl<DeptPlan, DeptPlanI
 		if(logger.isDebugEnabled()) logger.debug("加载部门计划类型［"+ type +"］名称...");
 		if(type == null  || this.typeMap == null || this.typeMap.size() == 0) return null;
 		return this.typeMap.get(type);
-	}
-	/*
-	 * 更新计划成员。
-	 * @see com.examw.oa.service.plan.IDeptPlanService#updateDeptPlanMember(java.lang.String, com.examw.oa.model.plan.DeptPlanMemberInfo)
-	 */
-	@Override
-	public DeptPlanMemberInfo updateDeptPlanMember(String deptPlanId, DeptPlanMemberInfo info) {
-		if(logger.isDebugEnabled())logger.debug("更新计划成员数据...");
-		if(info == null) return null;
-		Boolean isAdded = false;
-		DeptPlanMember  data = StringUtils.isEmpty(info.getId()) ? null : this.deptPlanMenberDao.load(DeptPlanMember.class, info.getId());
-		if(isAdded = (data == null)){
-			if(StringUtils.isEmpty(info.getId())){
-				info.setId(UUID.randomUUID().toString());
-				info.setCreateTime(new Date());
-			}
-			data = new DeptPlanMember();
-		}
-		if(!isAdded)info.setCreateTime(data.getCreateTime());
-		BeanUtils.copyProperties(info, data);
-		if(!StringUtils.isEmpty(info.getEntryId()) && (data.getEntry() == null || !data.getEntry().getId().equalsIgnoreCase(info.getEntryId()))){
-			Entry e = this.entryDao.load(Entry.class, info.getEntryId());
-			if(e != null) data.setEntry(e);
-		}
-		if(!StringUtils.isEmpty(info.getEmplId()) && (data.getEmployee() == null || !data.getEmployee().getId().equalsIgnoreCase(info.getEntryId()))){
-			Employee e = this.employeeDao.load(Employee.class, info.getEmplId());
-			if(e != null) data.setEmployee(e);
-		}
-		if(data.getEntry() != null) info.setEntryName(data.getEntry().getName());
-		if(data.getEmployee() != null) info.setEmplName(data.getEmployee().getName());
-		DeptPlan plan=this.deptPlanDao.load(DeptPlan.class, deptPlanId);
-		if(plan !=null){
-			info.setDeptPlanId(deptPlanId);
-		}
-		if(isAdded)this.deptPlanMenberDao.save(data);
-		return info;
 	}
 }
